First attempt at a ParameterTranspiler
This class implements a first pass at traversing the AST in order to find where
the WorkflowParams object is defined within the scargo Python script.
Finally, the transpiler converts the WorkflowParams arguments back into
a Python dictionary and then writes it to YAML file in the same
directory as the scargo Python script. Also added a custom
`ScargoTranspilerError` which is supposed to make the whole scargo
experience more fun!